refactor(composebinary): move to part of libstac
