+++ libstack/composebinary/utils_test.go
@@ -1,4 +1,4 @@
-package wrapper
+package composebinary
 
 import (
 	"testing"
